+++ core/data/src/main/java/kr/co/data/source/remote/DiaryRemoteDataSource.kt
@@ -12,7 +12,7 @@ interface DiaryRemoteDataSource {
         month: String
     ): Flow<List<DiaryData>>
 
-    suspend fun fetchDetail(id: Int): DiaryData
+    suspend fun fetchDetail(id: Long): DiaryData
 
     suspend fun searchList(
         crop: String,
@@ -22,7 +22,7 @@ interface DiaryRemoteDataSource {
     ): Flow<List<DiaryData>>
 
     suspend fun update(
-        id: Int,
+        id: Long,
         crop: String,
         date: String,
         holidayList: List<HolidayData>,
@@ -34,7 +34,7 @@ interface DiaryRemoteDataSource {
         memo: String
     )
 
-    suspend fun delete(id: Int)
+    suspend fun delete(id: Long)
 
     suspend fun create(
         crop: String,
